Added Sprite Sheets for Last cutscene Mike Pt 2
Added five sprite sheets for the last cutscene in Mike part 2.  As labeled, the first three specified around mike are animations for him saving stickman, him carrying and running with stickman, and then jumping down the hole, which play one right after the other.  The last 2 are the animations for the enemies coming down on Mike, having to be split into 2 parts due to the resolution cap, but play right after the other