finished 4.1 Mike Fight Scribble cutscence (#14)
* finished 4.1 Mike Fight Scribble Cutscence

* finished 4.1 Mike Fight Scribble cutscence

* 4.2 mike saved cutscene

can't figure out the orange object to disappear in the beginning tho! committing it bcs imma be gone for thanksgiving so just in case you guys wanna see the progress. it's there!

thanks :)

* 4.2 & 4.3 cutscene

4.2 fix & create 4.3 cutscenes

* Update Mike Part 1 - Copy Natasha.unity

* finished 4.1 Mike Fight Scribble Cutscence

* finished 4.1 Mike Fight Scribble cutscence

* 4.2 mike saved cutscene

can't figure out the orange object to disappear in the beginning tho! committing it bcs imma be gone for thanksgiving so just in case you guys wanna see the progress. it's there!

thanks :)

* 4.2 & 4.3 cutscene

4.2 fix & create 4.3 cutscenes

* Update Mike Part 1 - Copy Natasha.unity

* fix cutscenes

* fix gitignore

* fix committed merge conflicts

---------

Co-authored-by: someguy13 <ryan.kwong@gmail.com>
Co-authored-by: jalenng <jalen_ng@outlook.com>