Update mike voice (#152)
* New Mike Voice Clips

Old Mike Voice didn't set the tone of the cutscenes, so this is for his new voice. Also, removed the wrapUp() because it's not needed and causes a clicking sound at the end of every sentence.

* New Mike Voice Clips

Old Mike Voice didn't set the tone of the cutscenes, so this is for his new voice. Also, removed the wrapUp() because it's not needed and causes a clicking sound at the end of every sentence.

* Mike Voice Update

Added new clip, removed two that didn't fit.