Wall respawn based on checkpoints
- Removed all that calculation from wall respawn
- New script added onto checkpoints (in chase lvl only) that will update wall's respawn position when triggered.

(Only note: The checkpoint in the shortcut doesn't have this script, because its pretty funky with the way scribble wall paths. TLDR, if you set a new respawnpos for the shortcut, the wall navigates backward, thru the points it normally hits in its path. This is only a problem for the shortcut checkpoint, which is out of the normal path the player and wall take.
I think this problem should be solved by implementing another feature I plan on, catch up for the scribble wall (increasing speed based on distance from player)