Scribble wall softlock fix (#195)
* More respawn logic!!

- Removed unnecessary scripts from old scribble wall logic
- Deleted respawn points in chase level
- scribble wall now factors in player's distance from their respawn to make sure it spawns in behind them (needs testing)

* Logic Updates for Scribble Wall

- Respawns upon all player deaths now
- Deleted collision(), not needed now
      Optimization B)
- Commented StartRespawn(), formatted a little
- Getter function in PlayerMovement for isDead

* Fine Tuning for Scribble Wall

- Added more checkpoints for less awkwardness (idk howt o describe it)
- Moved some pathpoints around
- RespawnOffset set to 15
- Now will detect whether path is moving horizontally or vertically; Based on that, the wall's respawn will detect the player's vertical or horizontal distance from their checkpoints

* Wall respawn based on checkpoints

- Removed all that calculation from wall respawn
- New script added onto checkpoints (in chase lvl only) that will update wall's respawn position when triggered.

(Only note: The checkpoint in the shortcut doesn't have this script, because its pretty funky with the way scribble wall paths. TLDR, if you set a new respawnpos for the shortcut, the wall navigates backward, thru the points it normally hits in its path. This is only a problem for the shortcut checkpoint, which is out of the normal path the player and wall take.
I think this problem should be solved by implementing another feature I plan on, catch up for the scribble wall (increasing speed based on distance from player)

* merge main

---------

Co-authored-by: jalenng <jalen_ng@outlook.com>